requirements.txt
fastapi
uvicorn
python-telegram-bot
google-generativeai
python-dotenv